Begin refactor: Remove Room struct
to be replaced my Mission. There was no need for that extra layer
